Break long combinatorial paths through PerformanceMonitor
The input to PerformanceMonitor's send_performance_events included the
read port of an mkDWireOR in each cache collecting all the events,
causing send_performance_events to have all the cache logic upstream
from it. Since send_performance_events < read_counters, this then in
turn goes on to cause CSR logic to be dependent on all this. It's a
miracle there were no scheduling conflicts.

Instead, delay reporting the caches' events until the start of the next
cycle, since this is how PerformanceMonitor is expecting to be used (and
how Toooba uses it). We could register the events in CPU rather than the
caches, but this follows the Toooba design of using mkDRegOR to merge
the registering with the mkDWireOR.

Also register AXI and external (tag controller) events; whilst these
don't show up on the critical path, it's possible they'd become an issue
in future, and an extra cycle should not be an issue for either of them,
especially the already-asynchronous external events.

Hopefully this catches all the paths of this nature.